move tests out, move MaybeTimeOfFlight
